- Refactored network module from java to Android lib - Added network interceptor to ServiceGenerator - Organization of versions and dependencies in a single point Andromeda.gradle
