+++ app/build.gradle
@@ -23,37 +23,32 @@ android {
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
     // Support lib compat
-    implementation "com.android.support:recyclerview-v7:$support_version"
-    implementation "com.android.support:appcompat-v7:$support_version"
-    implementation "com.android.support:cardview-v7:$support_version"
-    implementation "com.android.support:design:$support_version"
-    implementation "com.android.support:support-v4:$support_version"
-    implementation "com.android.support:support-annotations:$support_version"
-
-    implementation"org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
+    implementation "com.android.support:recyclerview-v7:${versions.support_lib}"
+    implementation "com.android.support:appcompat-v7:${versions.support_lib}"
+    implementation "com.android.support:cardview-v7:${versions.support_lib}"
+    implementation "com.android.support:design:${versions.support_lib}"
+    implementation "com.android.support:support-v4:${versions.support_lib}"
+    implementation "com.android.support:support-annotations:${versions.support_lib}"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jre7:${versions.kotlin}"
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
     //youtube
     implementation files('libs/YouTubeAndroidPlayerApi.jar')
-
     // Architecture Components
-    implementation "android.arch.persistence.room:runtime:$rootProject.roomVersion"
-    annotationProcessor "android.arch.persistence.room:compiler:$rootProject.roomVersion"
-    implementation "android.arch.lifecycle:extensions:$archLifecycleVersion"
-    annotationProcessor "android.arch.lifecycle:compiler:$archLifecycleVersion"
-
+    implementation "android.arch.persistence.room:runtime:${versions.room}"
+    annotationProcessor "android.arch.persistence.room:compiler:${versions.room}"
+    implementation "android.arch.lifecycle:extensions:${versions.archLifecycle}"
+    annotationProcessor "android.arch.lifecycle:compiler:${versions.archLifecycle}"
     //Network
-    compile project(path: ':network-client')
-    //TODO figure out how to remove retorfit dependencies
-    implementation 'com.squareup.retrofit2:retrofit:2.4.0'
-    implementation 'com.squareup.retrofit2:converter-gson:2.4.0'
-
+    implementation project(path: ':network')
+    //TODO figure out how to handle common dependencies among modules
+    implementation deps.retrofit2
+    implementation "com.squareup.retrofit2:converter-gson:${versions.retrofit}"
+    implementation deps.stetho
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-
     //Utilities
     implementation 'com.nostra13.universalimageloader:universal-image-loader:1.9.5'
-
 }
