+++ app/src/main/java/demos/expmind/andromeda/AndromedaApp.kt
@@ -1,9 +1,9 @@
 package demos.expmind.andromeda
 
 import android.app.Application
+import com.facebook.stetho.Stetho
 import com.nostra13.universalimageloader.core.ImageLoader
 import com.nostra13.universalimageloader.core.ImageLoaderConfiguration
-import demos.expmind.network_client.ServiceGenerator
 
 /**
  * Created by RAJ1GA on 02/10/2018.
@@ -14,5 +14,11 @@ class AndromedaApp : Application() {
         super.onCreate()
         //Init Image loader
         ImageLoader.getInstance().init(ImageLoaderConfiguration.createDefault(this))
+        if (BuildConfig.DEBUG) {
+            Stetho.initialize(Stetho.newInitializerBuilder(this)
+                    .enableDumpapp(Stetho.defaultDumperPluginsProvider(this))
+                    .enableWebKitInspector(Stetho.defaultInspectorModulesProvider(this))
+                    .build())
+        }
     }
 }
