+++ build.gradle
@@ -1,19 +1,29 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext {
-        kotlin_version = '1.1.51'
-        roomVersion = '1.1.1'
-        archLifecycleVersion = '1.1.1'
-        support_version = '27.1.1'
-    }
+
+    ext.versions = [
+            kotlin : '1.1.51',
+            room : '1.1.1',
+            archLifecycle : '1.1.1',
+            support_lib : '27.1.1',
+            retrofit : '2.4.0',
+            stetho : '1.5.0'
+    ]
+
+    ext.deps = [
+        retrofit2 : "com.squareup.retrofit2:retrofit:${versions.retrofit}",
+        stetho : "com.facebook.stetho:stetho:${versions.stetho}",
+        networkInterceptor : "com.facebook.stetho:stetho-okhttp3:${versions.stetho}"
+
+    ]
     repositories {
         google()
         jcenter()
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:3.0.1'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:${versions.kotlin}"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
