--- network-client/build.gradle
@@ -1,14 +1,0 @@
-apply plugin: 'java-library'
-
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.squareup.retrofit2:retrofit:2.4.0'
-    implementation 'com.squareup.retrofit2:converter-gson:2.4.0'
-    // Network traffic analyzers
-    implementation 'com.squareup.okhttp3:logging-interceptor:3.8.0'
-    compile 'com.facebook.stetho:stetho:1.5.0'
-    compile 'com.facebook.stetho:stetho-okhttp3:1.5.0'
-}
-
-sourceCompatibility = "1.7"
-targetCompatibility = "1.7"
